separa functions e lista fetch do banco
