+++ components/Home.js
@@ -1,86 +1,25 @@
-import { StyleSheet, View, SafeAreaView, ImageBackground, Text } from 'react-native'
+import { StyleSheet, View, SafeAreaView, ImageBackground, Text, ScrollView } from 'react-native'
 import React from 'react'
 
 const Home = () => {
-
-    const [horario, setaHorario] = React.useState('');
-    const [vagas, setaVagas] = React.useState('');
-    const [enderecoInicial, setaEnderecoInicia] = React.useState('');
-    const [enderecoFinal, setaEnderecoFnicia] = React.useState('');
+    const [lista, setaLista] = React.useState([]);
 
     const endpoint = 'https://github-sq75eata2q-uc.a.run.app/api/v1/carona';
 
-    const criarCarona = () => {     
-    
-      let carona = {
-        horario: horario,
-        vagas: vagas,
-        enderecoInicial: enderecoInicial,
-        enderecoFinal: enderecoFinal,          
-      }
-      
-      let exp = new RegExp('^([01]\d|2[0-3]):([0-5]\d):([0-5]\d)$');
-      let horarioValido = carona.horario.match(exp) > 0; 
-     
-      if (horarioValido) {
-        fetch(endpoint, {
-            method: "POST",
-            body: JSON.stringify(carona),
-            headers: {"Content-type": "application/json; charset=UTF-8"}
-        })
-        .then(response => response.json()) 
-        .then(json => console.log(json))
-        .catch(err => console.log(err))
- 
-      } else {
-          console.log(`Favor, inserir horário válido no formato HH:MM:SS`)
-          Alert.alert(`Favor, inserir horário válido no formato HH:MM:SS`);
-      }            
-    }
+    React.useEffect(() => {
+        listarCaronaAll();   
+    }, []);
 
     const listarCaronaAll = () => {     
-     
       fetch(endpoint, {
         method: "GET",
         headers: {"Content-type": "application/json; charset=UTF-8"}
       })
-      .then(response => response.json()) 
-      .then(json => console.log(json))
-      .catch(err => console.log(err))      
-    }
-
-    const listarCaronaById = (id) => {      
-      
-      fetch(`${endpoint}/${id}`, {
-        method: "GET",
-        headers: {"Content-type": "application/json; charset=UTF-8"}
-      })
-      .then(response => console.log(response))
-      .catch(err => console.log(err))   
-    }
-
-    const atualizarCaronaById = (id) => {          
-      
-      fetch(`${endpoint}/${id}`, {
-        method: "PUT",
-        body: JSON.stringify(carona),
-        headers: {"Content-type": "application/json; charset=UTF-8"}
-      })
-      .then(response => response.json()) 
-      .then(json => console.log(json))
+      .then(response => response.json())
+      .then(json => setaLista(json))
       .catch(err => console.log(err))
     }
-
-    const deletarCaronaById = (id) => {    
-
-      fetch(`${endpoint}/${id}`, {
-        method: "DELETE",
-        headers: {"Content-type": "application/json; charset=UTF-8"}
-      })
-      .then(response => response.json()) 
-      .then(json => console.log(json))
-      .catch(err => console.log(err))
-    }        
+        
 
     return (
         <SafeAreaView style={styles.container}>
@@ -89,8 +28,16 @@ const Home = () => {
             </View>
             <ImageBackground source={require('../assets/images/logo.png')} style={styles.ridesList} imageStyle={styles.image}>
                 <Text style={styles.openRidesTitle}>Caronas abertas</Text>
-                <Text style={styles.ride}>objeto de uma lista aqui!</Text>
-    
+                <ScrollView>
+                    {lista.map(carona => (
+                        <View key={carona.id}>
+                            <Text style={styles.ride}>Horário: {carona.horario}</Text>
+                            <Text style={styles.ride}>Número de vagas: {carona.vagas}</Text>
+                            <Text style={styles.ride}>Ponto de partida: {carona.enderecoInicial}</Text>
+                            <Text style={styles.ride}>Destino: {carona.enderecoFinal}</Text>
+                        </View>
+                    ))}
+                </ScrollView>
             </ImageBackground>
         </SafeAreaView>    
     )
